requirements/pip-install.txt
#
# gpip packages required when installing this tool (runtime requirements)
#
awscli>=1.11.0,<1.12.0
boto3>=-1.4.0,<1.5.0
google-api-python-client>=1.6.0,<1.7.0
#google-cloud
jinja2>=2.9.0,<2.10.0
Mako>=1.0.0,<1.1.0

PyYAML==3.12

requests>=2.13.0,<2.20.0
six
